third_party/blink/web_tests/svg/dynamic-updates/SVGFETurbulenceElement-svgdom-numOctaves-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'numOctaves' property of the SVGFETurbulenceElment object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS turbulence.numOctaves.baseVal is 1
PASS turbulence.numOctaves.baseVal is 3
PASS successfullyParsed is true

TEST COMPLETE
